Extract CLI token and task compute helpers
